Dockerで動かないのを修正する
環境が変わった？多少修正が必要だった。

  * Macで実行すると、Ctrlを押さないとキーが入力されない
    * なんかのタイミングで普通に入力できるようになる
  * `dir` とすると、 `General Exception` で止まってしまう

+++ apilib/apisrc/strlen.c
@@ -1,7 +1,7 @@
 #include "stdio.h"
 
-int strlen(const char* str) {
-  int len = 0;
+size_t strlen(const char* str) {
+  size_t len = 0;
   for (; *str != '\0'; ++str, ++len);
   return len;
 }
